Returned processor IDs to be int

+++ packages/panzer/adapters/stk/src/stk_interface/Panzer_STK_Interface.hpp
@@ -48,7 +48,7 @@ buildElementDescriptor(stk::mesh::EntityId elmtId,std::vector<stk::mesh::EntityI
 
 class STK_Interface {
 public:
-   typedef double ProcIdData;
+   typedef int ProcIdData;
    typedef stk::mesh::Field<double> SolutionFieldType;
    typedef stk::mesh::Field<double,stk::mesh::Cartesian> VectorFieldType;
    typedef stk::mesh::Field<ProcIdData> ProcIdFieldType;
